Moved TeamPolicy CTAD tests into its own file and now build it under COMPILE_ONLY_SOURCES
