+++ core/unit_test/TestTeamPolicyConstructors.hpp
@@ -169,127 +169,4 @@ TEST(TEST_CATEGORY, team_policy_runtime_parameters) {
   // clang-format on
 }
 
-struct TestTeamPolicyCTADS {
-  struct SomeExecutionSpace {
-    using execution_space = SomeExecutionSpace;
-    using size_type       = size_t;
-  };
-  static_assert(Kokkos::is_execution_space_v<SomeExecutionSpace>);
-
-  struct ImplicitlyConvertibleToDefaultExecutionSpace {
-    operator Kokkos::DefaultExecutionSpace() const;
-  };
-  static_assert(!Kokkos::is_execution_space_v<
-                ImplicitlyConvertibleToDefaultExecutionSpace>);
-
-  static inline Kokkos::DefaultExecutionSpace des;
-  static inline ImplicitlyConvertibleToDefaultExecutionSpace notEs;
-  static inline SomeExecutionSpace ses;
-  static inline TEST_EXECSPACE es;
-
-  using TeamPolicyTestExecSpace = std::conditional_t<
-      std::is_same_v<TEST_EXECSPACE, Kokkos::DefaultExecutionSpace>,
-      Kokkos::TeamPolicy<>, Kokkos::TeamPolicy<TEST_EXECSPACE>>;
-
-  static inline int i;
-
-  // Default construction deduces to TeamPolicy<>
-  static_assert(
-      std::is_same_v<Kokkos::TeamPolicy<>, decltype(Kokkos::TeamPolicy{})>);
-
-  // Execution space not provided deduces to TeamPolicy<>
-
-  static_assert(
-      std::is_same_v<Kokkos::TeamPolicy<>, decltype(Kokkos::TeamPolicy(i, i))>);
-  static_assert(std::is_same_v<Kokkos::TeamPolicy<>,
-                               decltype(Kokkos::TeamPolicy(i, i, i))>);
-  static_assert(std::is_same_v<Kokkos::TeamPolicy<>,
-                               decltype(Kokkos::TeamPolicy(i, Kokkos::AUTO))>);
-  static_assert(
-      std::is_same_v<Kokkos::TeamPolicy<>,
-                     decltype(Kokkos::TeamPolicy(i, Kokkos::AUTO, i))>);
-  static_assert(std::is_same_v<Kokkos::TeamPolicy<>,
-                               decltype(Kokkos::TeamPolicy(i, Kokkos::AUTO,
-                                                           Kokkos::AUTO))>);
-  static_assert(
-      std::is_same_v<Kokkos::TeamPolicy<>,
-                     decltype(Kokkos::TeamPolicy(i, i, Kokkos::AUTO))>);
-
-  // DefaultExecutionSpace deduces to TeamPolicy<>
-
-  static_assert(std::is_same_v<Kokkos::TeamPolicy<>,
-                               decltype(Kokkos::TeamPolicy(des, i, i))>);
-  static_assert(std::is_same_v<Kokkos::TeamPolicy<>,
-                               decltype(Kokkos::TeamPolicy(des, i, i, i))>);
-  static_assert(
-      std::is_same_v<Kokkos::TeamPolicy<>,
-                     decltype(Kokkos::TeamPolicy(des, i, Kokkos::AUTO))>);
-  static_assert(
-      std::is_same_v<Kokkos::TeamPolicy<>,
-                     decltype(Kokkos::TeamPolicy(des, i, Kokkos::AUTO, i))>);
-  static_assert(std::is_same_v<Kokkos::TeamPolicy<>,
-                               decltype(Kokkos::TeamPolicy(des, i, Kokkos::AUTO,
-                                                           Kokkos::AUTO))>);
-  static_assert(
-      std::is_same_v<Kokkos::TeamPolicy<>,
-                     decltype(Kokkos::TeamPolicy(des, i, i, Kokkos::AUTO))>);
-
-  // Convertible to DefaultExecutionSpace deduces to TeamPolicy<>
-
-  static_assert(std::is_same_v<Kokkos::TeamPolicy<>,
-                               decltype(Kokkos::TeamPolicy(notEs, i, i))>);
-  static_assert(std::is_same_v<Kokkos::TeamPolicy<>,
-                               decltype(Kokkos::TeamPolicy(notEs, i, i, i))>);
-  static_assert(
-      std::is_same_v<Kokkos::TeamPolicy<>,
-                     decltype(Kokkos::TeamPolicy(notEs, i, Kokkos::AUTO))>);
-  static_assert(
-      std::is_same_v<Kokkos::TeamPolicy<>,
-                     decltype(Kokkos::TeamPolicy(notEs, i, Kokkos::AUTO, i))>);
-  static_assert(std::is_same_v<Kokkos::TeamPolicy<>,
-                               decltype(Kokkos::TeamPolicy(
-                                   notEs, i, Kokkos::AUTO, Kokkos::AUTO))>);
-  static_assert(
-      std::is_same_v<Kokkos::TeamPolicy<>,
-                     decltype(Kokkos::TeamPolicy(notEs, i, i, Kokkos::AUTO))>);
-
-  // SES != DefaultExecutionSpace deduces to TeamPolicy<SES>
-
-  static_assert(std::is_same_v<Kokkos::TeamPolicy<SomeExecutionSpace>,
-                               decltype(Kokkos::TeamPolicy(ses, i, i))>);
-  static_assert(std::is_same_v<Kokkos::TeamPolicy<SomeExecutionSpace>,
-                               decltype(Kokkos::TeamPolicy(ses, i, i, i))>);
-  static_assert(
-      std::is_same_v<Kokkos::TeamPolicy<SomeExecutionSpace>,
-                     decltype(Kokkos::TeamPolicy(ses, i, Kokkos::AUTO))>);
-  static_assert(
-      std::is_same_v<Kokkos::TeamPolicy<SomeExecutionSpace>,
-                     decltype(Kokkos::TeamPolicy(ses, i, Kokkos::AUTO, i))>);
-  static_assert(std::is_same_v<Kokkos::TeamPolicy<SomeExecutionSpace>,
-                               decltype(Kokkos::TeamPolicy(ses, i, Kokkos::AUTO,
-                                                           Kokkos::AUTO))>);
-  static_assert(
-      std::is_same_v<Kokkos::TeamPolicy<SomeExecutionSpace>,
-                     decltype(Kokkos::TeamPolicy(ses, i, i, Kokkos::AUTO))>);
-
-  // TEST_EXECSPACE deduces to TeamPolicyTestExecSpace
-
-  static_assert(std::is_same_v<TeamPolicyTestExecSpace,
-                               decltype(Kokkos::TeamPolicy(es, i, i))>);
-  static_assert(std::is_same_v<TeamPolicyTestExecSpace,
-                               decltype(Kokkos::TeamPolicy(es, i, i, i))>);
-  static_assert(
-      std::is_same_v<TeamPolicyTestExecSpace,
-                     decltype(Kokkos::TeamPolicy(es, i, Kokkos::AUTO))>);
-  static_assert(
-      std::is_same_v<TeamPolicyTestExecSpace,
-                     decltype(Kokkos::TeamPolicy(es, i, Kokkos::AUTO, i))>);
-  static_assert(std::is_same_v<TeamPolicyTestExecSpace,
-                               decltype(Kokkos::TeamPolicy(es, i, Kokkos::AUTO,
-                                                           Kokkos::AUTO))>);
-  static_assert(
-      std::is_same_v<TeamPolicyTestExecSpace,
-                     decltype(Kokkos::TeamPolicy(es, i, i, Kokkos::AUTO))>);
-};
-
 }  // namespace
